requirements.txt
openai
camoufox[geoip]
mcp[cli]
python-telegram-bot  # optional: for --telegram mode
